Fixed type errors after upgrading typescript

+++ src/app/lib/skill-badge-actions.ts
@@ -27,10 +27,12 @@ export const updateSkillBadge = async (skillBadgeId: string, formData: FormData)
     // Sanitize rich text fields before saving to database
     const sanitizedData = sanitizeFormData(validatedData);
 
-    const oldSkillBadge = await prisma.skillBadge.findUnique({
+    const oldSkillBadge = await prisma.skillBadge.findUniqueOrThrow({
         where: { id: validatedSkillBadgeId },
     });
-    await deleteOldBlob(oldSkillBadge.image_url, sanitizedData.image_url);
+    if (oldSkillBadge.image_url)
+        await deleteOldBlob(oldSkillBadge.image_url, sanitizedData.image_url);
+
     await prisma.skillBadge.update({ where: { id: validatedSkillBadgeId }, data: sanitizedData });
     revalidateTag(GlobalConstants.SKILL_BADGES, "max");
 };
@@ -42,6 +44,9 @@ export const deleteSkillBadge = async (skillBadgeId: string): Promise<void> => {
     const deletedSkillBadge = await prisma.skillBadge.delete({
         where: { id: validatedSkillBadgeId },
     });
-    await deleteOldBlob(deletedSkillBadge.image_url);
+
+    if (deletedSkillBadge.image_url)
+        await deleteOldBlob(deletedSkillBadge.image_url);
+
     revalidateTag(GlobalConstants.SKILL_BADGES, "max");
 };
